requirements.txt
matplotlib~=3.5.2
requests~=2.27.1
beautifulsoup4~=4.11.1
python-dateutil~=2.8.2